# public/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Stubborn
  description: "Documentation générée automatiquement pour l'application Stubborn."
  contact:
    name: 'Support Stubborn'
    email: contact@lewebpluschouette.fr
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
paths:
...
  /register:
    post:
      tags:
        - Inscription
      summary: 'Inscrire un nouvel utilisateur'
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                  format: email
                plainPassword:
                  type: string
                  format: password
                confirmPassword:
                  type: string
                  format: password
              type: object
      responses:
        '201':
          description: 'Compte créé, e-mail de vérification envoyé'
        '400':
          description: 'Données invalides ou mots de passe non concordants'
...
  /verify/email:
    get:
      tags:
        - Inscription
      summary: "Valider le lien de vérification d'e-mail"
      operationId: verifyEmail
      parameters:
        -
          name: id
          in: query
          description: 'Identifiant du nouvel utilisateur'
          required: true
          schema:
            type: integer
      responses:
        '302':
          description: 'Redirection vers la page d’accueil après succès'
        '400':
          description: 'Lien invalide ou expiré'
  /reset-password:
    post:
      tags:
        - 'Mot de passe'
      summary: 'Demander un lien de réinitialisation'
      description: 'Display & process form to request a password reset.'
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                  format: email
              type: object
      responses:
        '202':
          description: 'E‑mail envoyé (ou réponse générique)'
        '400':
          description: 'Données invalides'
...
  '/reset-password/reset/{token}':
    post:
      tags:
        - 'Mot de passe'
      summary: 'Enregistrer le nouveau mot de passe'
      operationId: resetPassword
      parameters:
        -
          name: token
          in: path
          description: 'Jeton unique reçu par e‑mail'
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                password:
                  type: string
                  format: password
              type: object
      responses:
        '204':
          description: 'Mot de passe réinitialisé'
        '400':
          description: 'Jeton invalide ou expiré'
...